requirements.txt
django>=3.0,<4.0
django-ckeditor>=6.0.0
django-colorfield==0.4.1
gunicorn>=20.1.0
psycopg2-binary>=2.8
python-dotenv==0.16.0
secret-key-generator==0.0.7